Simplify menu construction code and fix positioning bug on first input
